pyproject.toml
[tool.poetry]
name = "py-web-hw14"
version = "0.1.0"
description = "hw11-hw12-hw13.1-hw14"
authors = ["Artem Dorofeev <130536391+artem-dorofeev@users.noreply.github.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.104.1"
uvicorn = "^0.24.0.post1"
sqlalchemy = "^2.0.23"
psycopg2 = "^2.9.9"
alembic = "^1.13.0"
jinja2 = "^3.1.2"
faker = "^20.1.0"
pydantic = {version = "2.4.2", extras = ["email"]}
python-jose = "^3.3.0"
passlib = "^1.7.4"
python-multipart = "^0.0.6"
libgravatar = "^1.0.4"
fastapi-mail = "^1.4.1"
python-dotenv = "^1.0.0"
redis = "4.2"
fastapi-limiter = "^0.1.5"
cloudinary = "^1.37.0"


[tool.poetry.group.dev.dependencies]
sphinx = "^7.2.6"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
